refactor: separate matchday API wrappers from pure resolver logic
Cleanup requested during review of the #818 changes - the matchday
service files mixed I/O (fetch+map from the matchday API) with pure
functions (filter/sort already-fetched data) in the same modules,
and matchdaySiteConfig.ts didn't fit its own folder's naming
convention (it reads Sanity, not the matchday API).

- getMatchdayClubId() moved into lib/content/siteSettings.ts, next to
  getSiteSettings() which it wraps - deleted matchdaySiteConfig.ts.
  leagueOptionService.ts now reuses it instead of re-extracting
  siteSettings.matchday.clubId itself.
- resolveMatchdayNextMatch/resolveMatchdayPreviousMatch (pure - no
  fetching) moved out of matchdayMatchService.ts into a new
  matchResolverService.ts. matchdayMatchService.ts now only does
  fetch+map (getMatchdayFixturesForLeague).

No behavior change - reverified fixture/next-match/previous-match/
table mapping against the real matchday API after the move.
